Allow for trigger-sound path to start with ~
Expand '~' in the trigger-sound argument so a user can pass in '~/test.wav'
Calling os.path.expanduser on a string that doesn't start with ~ results in a no-op.

Add a log message when the trigger-sound argument is present, but does not exist according to os.path.exists to help with debugging.