Remove button trigger from assistant_library_demo.
Apparently this is too complicated. Will make another demo to showcase
how to properly configure the button trigger with assistant library.

The reason is the assistant library blocks the main thread and prevents
the button GPIO driver from calling the on_button_pressed() callback. As
a result a separate thread is required for the assistant library event
loop.