Add assistant library demo. (#104)
* Add aiy.voicehat.get_status_ui

The function returns a new driver to access the status-led service. In
addition, this change adds example usage to assistant_grpc_demo.

Tested:
- Deployed to RPi3 + VoiceHat.
- Ran src/assistant_grpc_demo.py

* Add assistant library demo.

The demo is src/assistant_library_demo.py.

Tested:
- Deployed to RPi3 + VoiceHat.
- Ran src/assistant_library_demo.py

* Also press the button to start a conversation.

This change enabled the GPIO trigger (button) to be used with the
assistant_library_demo. Now the user can either say the hot word "Okay,
Google" or press the VoiceHat arcade button to start a conversation.

* Remove button trigger from assistant_library_demo.

Apparently this is too complicated. Will make another demo to showcase
how to properly configure the button trigger with assistant library.

The reason is the assistant library blocks the main thread and prevents
the button GPIO driver from calling the on_button_pressed() callback. As
a result a separate thread is required for the assistant library event
loop.