Flexible python API to control KTD LED driver.
Improving original RGBLED API. Blink pattern and channel state (always
on, always off, or following the pattern) are orthogonal and can be
configured separately. API is flexible enough and could be reused to
control custom LEDs attached to KTD driver.  Shortcuts are provided to
simplify control of out-of-the-box available RGB and Privacy LEDs. Check
typical API usage in leds_example.py.

Change-Id: I64ff1876ee0b943c735a72285239ba6d1f0e1e77