Also install google-assistant-library on armv6l
I'm running a raspberry pi zero w (arm6l). I've done all the setup steps. 
When running the example scripts, the `google-assistant-library` was missing. So I installed it the same way like in the install-deps.sh does it for ARMv7 and it works fine. So this I guess it's not only available on v7 but on v6?