Add config and RST files for Sphinx-generated Python reference.
Each .rst file is a descriptor for a page to generate. So there's
one for each major module. In cases such as aiy.assistant and
aiy.voice.models, I decided it's best to create a package-level
page that lists each sub-module on the same page.

The overall organization is not literal to the organization in the
source code. I've taken some inspiration from other Python projects
and decided to organize the reference in a way that should be more
intuitive to users who own a specific kit. So, there's a landing
page for Voice and Vision Kit, where I list the modules that are
intended for each one. I've simply reused the existing markdown pages
for each kit, reorganized a bit and added links to the respective
modules that are essential to each.

However, some of the "kit-specific" modules are not actually restricted
to those kits with hardware dependencies, so I've added descriptions
where appropriate to say so (for example, the assistant library is under
Voice Kit, but could be used without the Voice Bonnet).

Besides all the .rst files, I've added some module summaries in the .py
files, including some sample code for a couple of them. But that was
merely to explore some of the tool and add a little info... This CL
is not intended to fill in the code comments, but instead just get the
tooling setup and enable us to publish an min-viable python reference.

I've also added a custom CSS file to override some of the "read the docs"
theme's defaults styles, such as code font. We can extend this in the
future if we want to further match the product website.

There's much more documentation to add now that it's easy to see where
it's missing.

Change-Id: Ia735a313e0e031bc0bd5e51fa0bed34b00d7dd5b